+++ library/vibron/source/vibration/vibration.cpp
@@ -12,17 +12,17 @@ namespace vibron {
 void Vibration::construct_excitation() {
     excitation_ = 0;
     excited_modes_.clear();
-    for (size_t irred = 0; irred < phonons_.size(); irred++)
-    for (size_t mode = 0; mode < phonons_[irred].size(); mode++)
+    for (uint16_t irred = 0; irred < phonons_.size(); irred++)
+    for (uint16_t mode = 0; mode < phonons_[irred].size(); mode++)
     if (phonons_[irred][mode] > 0) {
         excitation_++;
-        excited_modes_.push_back(std::pair<size_t, size_t>(irred, mode));
+        excited_modes_.push_back(std::pair<uint16_t, uint16_t>(irred, mode));
     }
     excited_modes_.shrink_to_fit();
 }
 
 Vibration::Vibration() {}
-Vibration::Vibration(const std::vector<std::vector<size_t>> & _phonons)
+Vibration::Vibration(const std::vector<std::vector<uint16_t>> & _phonons)
 : phonons_(_phonons) {this->construct_excitation();}
 Vibration::Vibration(const std::vector<std::string> & lines) {
     phonons_.resize(lines.size());
@@ -36,9 +36,9 @@ Vibration::Vibration(const std::vector<std::string> & lines) {
 }
 Vibration::~Vibration() {}
 
-const std::vector<std::vector<size_t>> & Vibration::phonons() const {return phonons_;}
-const size_t & Vibration::excitation() const {return excitation_;}
-const std::vector<std::pair<size_t, size_t>> & Vibration::excited_modes() const {return excited_modes_;}
+const std::vector<std::vector<uint16_t>> & Vibration::phonons() const {return phonons_;}
+const uint16_t & Vibration::excitation() const {return excitation_;}
+const std::vector<std::pair<uint16_t, uint16_t>> & Vibration::excited_modes() const {return excited_modes_;}
 
 void Vibration::pretty_print(std::ostream & stream) const {
     for (size_t i = 0; i < phonons_.size(); i++) {
